refactor(main): session persistence helper を分離
Diff viewer の window/per-PR session 保存 helper を app::diff_viewer::session へ移し、main.rs を composition root に近づける。

Refs: #224
