+++ src/app/diff_viewer/mod.rs
@@ -5,6 +5,7 @@
 
 pub(crate) mod snapshot;
 pub(crate) mod refresh;
+pub(crate) mod session;
 pub(crate) mod state;
 pub(crate) mod toast;
 pub(crate) mod util;
